modify int32 to int64

+++ Re-Abyss/app/datastores/Migration/MigrationDataStore.cpp
@@ -6,19 +6,19 @@ namespace abyss::User
     MigrationDataStore::MigrationDataStore()
     {
     }
-    s3d::Array<s3d::int32> MigrationDataStore::versionIds() const
+    s3d::Array<s3d::int64> MigrationDataStore::versionIds() const
     {
         return Migration::MigrationHundler::Versions();
     }
-    s3d::Array<s3d::int32> MigrationDataStore::updateVersionIds(s3d::int32 currentVersionId) const
+    s3d::Array<s3d::int64> MigrationDataStore::updateVersionIds(s3d::int64 currentVersionId) const
     {
         return Migration::MigrationHundler::UpdateVersions(currentVersionId);
     }
-    void MigrationDataStore::up(s3d::int32 versionId) const
+    void MigrationDataStore::up(s3d::int64 versionId) const
     {
         Migration::MigrationHundler::Get(versionId)->up();
     }
-    void MigrationDataStore::down(s3d::int32 versionId) const
+    void MigrationDataStore::down(s3d::int64 versionId) const
     {
         Migration::MigrationHundler::Get(versionId)->down();
     }
