+++ Re-Abyss/app/datastores/MigrationVersion/base/IMigrationVersionDataStore.hpp
@@ -10,8 +10,8 @@ namespace abyss::User
 
         virtual bool create() const = 0;
 
-        virtual s3d::int32 selectCurrentVersion()const = 0;
-        virtual bool update(const s3d::Array<s3d::int32>& versions)const = 0;
-        virtual bool erase(s3d::int32 versionId)const = 0;
+        virtual s3d::int64 selectCurrentVersion()const = 0;
+        virtual bool update(const s3d::Array<s3d::int64>& versions)const = 0;
+        virtual bool erase(s3d::int64 versionId)const = 0;
     };
 }
