+++ Re-Abyss/app/utils/Migration/Migration.cpp
@@ -2,17 +2,17 @@
 
 namespace abyss::Migration
 {
-    s3d::Array<s3d::int32> MigrationHundler::Versions()
+    s3d::Array<s3d::int64> MigrationHundler::Versions()
     {
-        s3d::Array<s3d::int32> ret;
+        s3d::Array<s3d::int64> ret;
         for (auto&& m : m_migrations) {
             ret << m.first;
         }
         return ret.sorted();
     }
-    s3d::Array<s3d::int32> MigrationHundler::UpdateVersions(s3d::int32 currentVersion)
+    s3d::Array<s3d::int64> MigrationHundler::UpdateVersions(s3d::int64 currentVersion)
     {
-        s3d::Array<s3d::int32> ret;
+        s3d::Array<s3d::int64> ret;
         for (auto&& m : m_migrations) {
             if (m.first > currentVersion) {
                 // 現在のバージョンより上
@@ -21,7 +21,7 @@ namespace abyss::Migration
         }
         return ret.sorted();
     }
-    IMigration* MigrationHundler::Get(s3d::int32 version)
+    IMigration* MigrationHundler::Get(s3d::int64 version)
     {
         return m_migrations[version];
     }
