+++ Re-Abyss/app/utils/Migration/Migration.hpp
@@ -13,31 +13,31 @@ namespace abyss::Migration
         virtual void down() const = 0;
     };
 
-    template<s3d::int32 Id>
+    template<s3d::int64 Id>
     struct Migration;
 
     class MigrationHundler
     {
     public:
-        static s3d::Array<s3d::int32> Versions();
-        static s3d::Array<s3d::int32> UpdateVersions(s3d::int32 currentVersion);
+        static s3d::Array<s3d::int64> Versions();
+        static s3d::Array<s3d::int64> UpdateVersions(s3d::int64 currentVersion);
 
-        template<s3d::int32 Id>
+        template<s3d::int64 Id>
         static void Regist(IMigration* migration)
         {
             m_migrations[Id] = migration;
         }
 
-        static IMigration* Get(s3d::int32 version);
+        static IMigration* Get(s3d::int64 version);
     private:
         MigrationHundler() = default;
         MigrationHundler(const MigrationHundler& other) = delete;
         void operator=(const MigrationHundler& other) = delete;
     private:
-        inline static s3d::HashTable<s3d::int32, IMigration*> m_migrations;
+        inline static s3d::HashTable<s3d::int64, IMigration*> m_migrations;
     };
 
-    template<s3d::int32 Id>
+    template<s3d::int64 Id>
     struct Migration : public IMigration
     {
     public:
@@ -49,7 +49,7 @@ namespace abyss::Migration
 
     namespace detail
     {
-        consteval s3d::int32 versionId(const char* path)
+        consteval s3d::int64 versionId(const char* path)
         {
             // filename取得
             const char* filename = path;
@@ -60,7 +60,7 @@ namespace abyss::Migration
             }
             // バージョン値取得
             const char* str = filename;
-            s3d::int32 value = 0;
+            s3d::int64 value = 0;
             while (char c = *str) {
                 if (c <= '9' && c >= '0') {
                     value = (c - '0') + value * 10;
